refactor: move type checking let
